requirements.txt
docker-py
openai>=1.12.0,<2.0.0
ddgs>=9.5.2,<10.0.0
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
websocket-client>=1.8.0,<2.0.0
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
python-multipart>=0.0.6,<1.0.0